蓝桥抱佛脚 day3

+++ AcWing/蓝桥杯C++AB组/第一讲 递归与递推/1209. 带分数.cpp
@@ -20,7 +20,7 @@ int n;  // 输入的正整数
 int res = 0;
 
 bool check(int c, int a) {
-    int b = c * n - c * a;
+    long long  b = (long long) c * n - c * a;  // 防止溢出
     if (a == 0 || b == 0 || c == 0)
         return false;
